+++ src/main/java/pl/slowikowski/demo/crud/exception/NotFoundException.java
@@ -1,7 +1,7 @@
 package pl.slowikowski.demo.crud.exception;
 
 public class NotFoundException extends RuntimeException {
-    public NotFoundException(int id, String className) {
+    public NotFoundException(Long id, String className) {
         super("Could not find " + className + " with id = " + id);
     }
 }
